mmc: core: Attempt to set bus width even if legacy speed
The Zynq7 has SDIO interface that isn't compliant with the eMMC
high-speed timing spec and can be disabled with `broken-mmc-highspeed`
quirk.

With the quirk, the MMC core never attempts to set the 4 bit wide
interface which does meet the timing requirements.

This patch attempts to set the bus width as a last ditch effort and
expects a `bus-width` specification in the device tree.